backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
pydantic==2.9.2
pydantic-settings==2.5.2
qdrant-client==1.11.3
python-multipart==0.0.9
litellm==1.48.0
cryptography==43.0.1
httpx==0.27.2
beautifulsoup4==4.12.3
openai==1.51.0
pdfplumber==0.11.4
python-docx==1.1.2
